DG-122 commit fixed.
