+++ src/main/java/org/micks/DiscGolfApplication/events/DiscGolfDataService.java
@@ -1,12 +1,14 @@
 package org.micks.DiscGolfApplication.events;
 
 import lombok.extern.slf4j.Slf4j;
-import org.apache.poi.ss.usermodel.*;
+import org.apache.poi.ss.usermodel.Cell;
+import org.apache.poi.ss.usermodel.Row;
+import org.apache.poi.ss.usermodel.Sheet;
+import org.apache.poi.ss.usermodel.Workbook;
 import org.apache.poi.xssf.usermodel.XSSFWorkbook;
 import org.micks.DiscGolfApplication.connection.DiscGolfDbConnection;
-import org.springframework.beans.factory.annotation.Autowired;
+import org.micks.DiscGolfApplication.exceptions.ImportException;
 import org.springframework.stereotype.Service;
-import org.springframework.web.multipart.MultipartFile;
 
 import java.io.ByteArrayOutputStream;
 import java.io.IOException;
@@ -15,6 +17,7 @@ import java.sql.Connection;
 import java.sql.PreparedStatement;
 import java.sql.ResultSet;
 import java.sql.SQLException;
+import java.text.ParseException;
 import java.text.SimpleDateFormat;
 import java.util.Date;
 import java.util.List;
@@ -24,9 +27,6 @@ import java.util.List;
 public class DiscGolfDataService {
 
     private final DiscGolfEventService discGolfEventService;
-
-    @Autowired
-    private DiscGolfDbConnection dbConnection;
 
     public DiscGolfDataService(DiscGolfEventService discGolfEventService) {
         this.discGolfEventService = discGolfEventService;
@@ -69,129 +69,64 @@ public class DiscGolfDataService {
             throw new RuntimeException("Failed to generate Excel file", e);
         }
     }
-    public void importEventsExcel(MultipartFile file) {
-        try (InputStream is = file.getInputStream()) {
-            Workbook workbook = new XSSFWorkbook(is);
-            Sheet sheet = workbook.getSheetAt(0);
 
-            int created = 0;
-            int updated = 0;
-
-            for (int i = 1; i <= sheet.getLastRowNum(); i++) {
-                Row row = sheet.getRow(i);
-                if (row == null) continue;
-
-                Date tournamentDate = getCellValueAsDate(row.getCell(1));
-                String pdga = getCellValueAsString(row.getCell(2));
-                String tournamentTitle = getCellValueAsString(row.getCell(3)).trim();
-                String region = getCellValueAsString(row.getCell(4));
-                String registration = getCellValueAsString(row.getCell(5));
-                String vacancies = getCellValueAsString(row.getCell(6));
-
-                if (tournamentTitle.isEmpty() || tournamentDate == null) {
-                    log.warn("Skipping row {} - missing title or date", i);
-                    continue;
-                }
-
-                DiscGolfEventDTO event = new DiscGolfEventDTO(
-                        null,
-                        tournamentDate,
-                        pdga,
-                        tournamentTitle,
-                        region,
-                        registration,
-                        vacancies
-                );
-
-                if (eventExistsByTitle(tournamentTitle)) {
-                    updateEventByTitle(tournamentTitle, event);
-                    updated++;
-                } else {
-                    discGolfEventService.createEvent(event);
-                    created++;
-                }
-            }
-
-            log.info("Import completed. Created: {}, Updated: {}", created, updated);
-
+    public void importEventsExcel(InputStream is) {
+        Workbook workbook;
+        try {
+            workbook = new XSSFWorkbook(is);
         } catch (IOException e) {
-            throw new RuntimeException("Failed to import Excel file", e);
+            throw new RuntimeException("Failed to load Excel file", e);
         }
-    }
+        Sheet sheet = workbook.getSheetAt(0);
 
-    private String getCellValueAsString(Cell cell) {
-        if (cell == null) return "";
+        int created = 0;
+        int updated = 0;
 
-        return switch (cell.getCellType()) {
-            case STRING -> cell.getStringCellValue();
-            case NUMERIC -> {
-                if (DateUtil.isCellDateFormatted(cell)) {
-                    yield cell.getDateCellValue().toString();
-                }
-                yield String.valueOf(cell.getNumericCellValue());
+        for (int i = 1; i <= sheet.getLastRowNum(); i++) {
+            Row row = sheet.getRow(i);
+            if (row == null) continue;
+
+            Date tournamentDate = getCellValueAsDate(row.getCell(1));
+            String pdga = row.getCell(2).getStringCellValue();
+            String tournamentTitle = row.getCell(3).getStringCellValue().trim();
+            String region = row.getCell(4).getStringCellValue();
+            String registration = row.getCell(5).getStringCellValue();
+            String vacancies = row.getCell(6).getStringCellValue();
+
+            if (tournamentTitle.isEmpty() || tournamentDate == null) {
+                log.warn("Skipping row {} - missing title or date", i);
+                continue;
             }
-            case BOOLEAN -> String.valueOf(cell.getBooleanCellValue());
-            default -> "";
-        };
+
+            DiscGolfEventDTO event = new DiscGolfEventDTO(
+                    null,
+                    tournamentDate,
+                    pdga,
+                    tournamentTitle,
+                    region,
+                    registration,
+                    vacancies
+            );
+
+            if (discGolfEventService.eventExistsByTitle(tournamentTitle)) {
+                discGolfEventService.updateEventByTitle(tournamentTitle, event);
+                updated++;
+            } else {
+                discGolfEventService.createEvent(event);
+                created++;
+            }
+        }
+
+        log.info("Import completed. Created: {}, Updated: {}", created, updated);
     }
 
     private Date getCellValueAsDate(Cell cell) {
-        if (cell == null) return null;
-
+        String dateStr = cell.getStringCellValue();
+        SimpleDateFormat sdf = new SimpleDateFormat("yyyy-MM-dd");
         try {
-            if (cell.getCellType() == CellType.NUMERIC && DateUtil.isCellDateFormatted(cell)) {
-                return cell.getDateCellValue();
-            } else if (cell.getCellType() == CellType.STRING) {
-                String dateStr = cell.getStringCellValue();
-                SimpleDateFormat sdf = new SimpleDateFormat("yyyy-MM-dd");
-                return sdf.parse(dateStr);
-            }
-        } catch (Exception e) {
-            log.error("Failed to parse date from cell", e);
-        }
-        return null;
-    }
-    public boolean eventExistsByTitle(String tournamentTitle) {
-        try (Connection connection = dbConnection.connect()) {
-            String query = "SELECT COUNT(*) FROM Events WHERE tournamentTitle = ? AND status != 'DELETED'";
-            PreparedStatement stmt = connection.prepareStatement(query);
-            stmt.setString(1, tournamentTitle);
-
-            ResultSet resultSet = stmt.executeQuery();
-            if (resultSet.next()) {
-                return resultSet.getInt(1) > 0;
-            }
-            return false;
-        } catch (SQLException e) {
-            log.error("Error checking if event exists: {}", tournamentTitle, e);
-            return false;
-        }
-    }
-
-    private void updateEventByTitle(String tournamentTitle, DiscGolfEventDTO event) {
-        try (Connection connection = dbConnection.connect()) {
-            String query = "UPDATE Events SET " +
-                    "tournamentDate = ?, " +
-                    "pdga = ?, " +
-                    "region = ?, " +
-                    "registration = ?, " +
-                    "vacancies = ? " +
-                    "WHERE tournamentTitle = ? AND status != 'DELETED'";
-
-            PreparedStatement stmt = connection.prepareStatement(query);
-            stmt.setDate(1, new java.sql.Date(event.getTournamentDate().getTime()));
-            stmt.setString(2, event.getPdga());
-            stmt.setString(3, event.getRegion());
-            stmt.setString(4, event.getRegistration());
-            stmt.setString(5, event.getVacancies());
-            stmt.setString(6, tournamentTitle);
-
-            stmt.executeUpdate();
-            log.info("Updated event: {}", tournamentTitle);
-
-        } catch (SQLException e) {
-            log.error("Error updating event: {}", tournamentTitle, e);
-            throw new RuntimeException("Failed to update event", e);
+            return sdf.parse(dateStr);
+        } catch (ParseException e) {
+            throw new ImportException("Cannot parse date: " + dateStr);
         }
     }
 }
