+++ src/main/java/org/micks/DiscGolfApplication/exceptions/ImportException.java
@@ -1,0 +1,7 @@
+package org.micks.DiscGolfApplication.exceptions;
+
+public class ImportException extends RuntimeException {
+    public ImportException(String message) {
+        super(message);
+    }
+}
